requirements.txt
appdirs==1.4.3
Django==1.11.2
django-auth-ldap3==0.9.6
cached-property==1.3.0
colorama==0.3.9
django-cors-headers==2.1.0
django-environ==0.4.3
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
ldap==1.0.2
packaging==16.8
psycopg2==2.7.1
pyasn1==0.2.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
docker==2.3.0
docker-compose==1.13.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
gunicorn==19.7.1
jsonschema==2.6.0
ldap3==2.2.4
PyJWT==1.5.0
pyldap==2.4.35.1
PyYAML==3.12
requests==2.11.1
texttable==0.8.8
websocket-client==0.40.0
